Add GitHub action for rebasing release branches
OpenEmbedded layers use separate branches for the Yocto releases. This
is neat, as it allows for recipe changes specific to a Yocto release.
However, manually maintaining these branches is annoying.

This commit adds a GitHub Action which automatically tries to rebase the
Yocto release branches on the master branch after push to the latter.